requirements.txt
flask==3.0.3
web3==6.19.0
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.32.3